# resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "El login exitoso de UserService devuelve un Colaborador con un campo \"token\".\nDicho token se debe incluir como header de las llamadas del resto de las operaciones de cada servicio.",
    "version": "1.0-SNAPSHOT",
    "title": "Sulfur backend",
    "contact": {
      "name": "rsperoni@magnesium.coop"
    }
  },
  "host": "sulfur.mgcoders.com",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/horas": {
      "get": {
        "tags": [
          "horas"
        ],
        "summary": "Get horas",
        "description": "",
        "operationId": "findAll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hora"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "horas"
        ],
        "summary": "Create hora",
        "description": "",
        "operationId": "create",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Hora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Hora"
            }
          },
          "400": {
            "description": "Objeto inválido"
          },
          "500": {
            "description": "Error interno"
          },
          "401": {
            "description": "No Autorizado"
          },
          "409": {
            "description": "Código o Id ya existe"
          }
        }
      }
    },
    "/horas/user/{id}/{fecha_ini}/{fecha_fin}": {
      "get": {
        "tags": [
          "horas"
        ],
        "summary": "Get horas por usuario y fechas",
        "description": "",
        "operationId": "findAllByColaborador",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "fecha_ini",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fecha_fin",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Hora"
              }
            }
          },
          "400": {
            "description": "Objeto inválido"
          },
          "401": {
            "description": "No Autorizado"
          }
        }
      }
    },
    "/horas/{id}": {
      "get": {
        "tags": [
          "horas"
        ],
        "summary": "Get hora",
        "description": "",
        "operationId": "find",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Hora"
            }
          },
          "404": {
            "description": "Id no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "horas"
        ],
        "summary": "Edit hora",
        "description": "",
        "operationId": "edit",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Hora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Hora"
            }
          },
          "304": {
            "description": "Error: objeto no modificado"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Cargo": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "nombre": {
          "type": "string"
        },
        "codigo": {
          "type": "string"
        },
        "precioHoraHistoria": {
          "type": "array",
          "uniqueItems": true,
          "items": {
            "$ref": "#/definitions/PrecioHora"
          }
        }
      }
    },
    "Proyecto": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "codigo": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
    "Hora": {
      "type": "object",
      "required": [
        "colaborador",
        "dia",
        "horaIn",
        "horaOut"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "dia": {
          "type": "string",
          "format": "date",
          "example": "23/01/2017"
        },
        "horaIn": {
          "type": "string",
          "format": "date-time",
          "example": "08:15"
        },
        "horaOut": {
          "type": "string",
          "format": "date-time",
          "example": "17:34"
        },
        "subtotal": {
          "type": "string",
          "format": "date-time",
          "example": "08:00",
          "readOnly": true
        },
        "colaborador": {
          "$ref": "#/definitions/Colaborador"
        },
        "horaDetalleList": {
          "type": "array",
          "uniqueItems": true,
          "items": {
            "$ref": "#/definitions/HoraDetalle"
          }
        },
        "completa": {
          "type": "boolean",
          "readOnly": true,
          "default": false
        },
        "subtotalDetalles": {
          "type": "string",
          "format": "date-time",
          "example": "08:00",
          "readOnly": true
        }
      }
    },
    "HoraDetalle": {
      "type": "object",
      "required": [
        "proyecto",
        "tipoTarea"
      ],
      "properties": {
        "proyecto": {
          "$ref": "#/definitions/Proyecto"
        },
        "tipoTarea": {
          "$ref": "#/definitions/TipoTarea"
        },
        "duracion": {
          "type": "string",
          "format": "date-time",
          "example": "17:34"
        }
      }
    },
...
    "PrecioHora": {
      "type": "object",
      "required": [
        "precioHora",
        "vigenciaDesde"
      ],
      "properties": {
        "precioHora": {
          "type": "number",
          "format": "double"
        },
        "vigenciaDesde": {
          "type": "string",
          "format": "date",
          "example": "23/01/2017"
        }
      }
    },
    "TipoTarea": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "codigo": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
    "Colaborador": {
      "type": "object",
      "required": [
        "email",
        "nombre",
        "role"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "email": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        },
        "cargo": {
          "$ref": "#/definitions/Cargo"
        },
        "role": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    }
  }
}